rename collectionIcons.tsx for real
