+++ pages/[user]/components/collections/EditCollectionRow.tsx
@@ -2,7 +2,7 @@ import { Paper, Group, Button, Text, Box, Tooltip } from "@mantine/core";
 import { IconMenu2, IconPencil } from "@tabler/icons";
 
 import PRIVACY_LEVELS from "components/collections/privacyLevels";
-import { COLLECTION_ICONS } from "components/collections/collectionsIcons";
+import { COLLECTION_ICONS } from "components/collections/collectionIcons";
 import { CardCollection } from "types/makotools";
 
 function EditCollectionRow({
